move policies to own constants and allow escaping on DEFAULT_POLICIES
